Added remember scroll progress feature
I added an onscroll event which stores in LocalStorage the current scrollY relatively to the last chapter header. But before this it gets the stored current chapter and relative scroll and scrolls to the required position after scroolling to the required chapter.
It may not be the best solution, but it's at least something. This feature is a must!
This contribution was made in my free time.